deduping some eval code/cleaning up pipeline
